refactor(web): share cancellable request deadlines
