+++ apps/web/src/lib/api.ts
@@ -86,3 +86,21 @@ export async function api<T>(path: string, init: RequestInit = {}): Promise<T> {
   const text = await response.text();
   return text ? (JSON.parse(text) as T) : (undefined as T);
 }
+
+export async function apiWithTimeout<T>(path: string, signal?: AbortSignal): Promise<T> {
+  if (!signal) return api<T>(path);
+  signal.throwIfAborted();
+  // Keep caller cancellation and the deadline without requiring AbortSignal.any.
+  const controller = new AbortController();
+  const timeout = AbortSignal.timeout(DEFAULT_FETCH_TIMEOUT_MS);
+  const onCancel = () => controller.abort(signal.reason);
+  const onTimeout = () => controller.abort(timeout.reason);
+  signal.addEventListener("abort", onCancel, { once: true });
+  timeout.addEventListener("abort", onTimeout, { once: true });
+  try {
+    return await api<T>(path, { signal: controller.signal });
+  } finally {
+    signal.removeEventListener("abort", onCancel);
+    timeout.removeEventListener("abort", onTimeout);
+  }
+}
