+++ docs/features/realtime.md
@@ -175,7 +175,8 @@ progress while retaining targetless, workspace-wide presence events.
   when it has a saved cursor. An expired session returns the chat or embedded
   view to sign-in; a workspace permission error remains distinct from global
   sign-out. Temporary failures retry with the same processed cursor, so they
-  cannot skip missed events.
+  cannot skip missed events. Validation retains the usual 30-second request
+  deadline and is cancelled when leaving the workspace.
 - On WebSocket connect and each live wake, the server pages durable events with
   a higher `cursor` until it reaches the visible tail captured for that drain. If replay is
   interrupted, the client can reconnect with the last cursor it actually
